c4_rust::symbol
Struct SymbolTable

pub struct SymbolTable { /* private fields */ }

Symbol table for managing variables and functions

Implementations

impl SymbolTable

pub fn new() -> Self

Create a new symbol table

pub fn add( &mut self, name: &str, class: TokenType, typ: Type, value: i64, ) -> usize

Add a symbol to the symbol table

Arguments

name - Symbol name

class - Symbol class (Fun, Glo, Loc, etc.)

typ - Symbol type (INT, CHAR, PTR)

value - Symbol value or address

Returns

The index of the added symbol

pub fn get(&self, name: &str) -> Option<&Symbol>

Get a symbol by name

Arguments

name - Symbol name

Returns

The symbol if found, or None

pub fn get_mut(&mut self, name: &str) -> Option<&mut Symbol>

Get a mutable reference to a symbol

Arguments

name - Symbol name

Returns

Mutable reference to the symbol if found, or None

pub fn get_by_index(&self, index: usize) -> Option<&Symbol>

Get a symbol by index

Arguments

index - Symbol index

Returns

The symbol at the given index, or None

pub fn get_by_index_mut(&mut self, index: usize) -> Option<&mut Symbol>

Get a mutable reference to a symbol by index

Arguments

index - Symbol index

Returns

Mutable reference to the symbol at the given index, or None

pub fn exists(&self, name: &str) -> bool

Check if a symbol exists

Arguments

name - Symbol name

Returns

True if the symbol exists, false otherwise

pub fn enter_scope(&mut self)

Enter a new scope level

pub fn exit_scope(&mut self)

Exit the current scope level

pub fn current_scope_level(&self) -> usize

Get the current scope level

pub fn len(&self) -> usize

Get the number of symbols in the table

pub fn is_empty(&self) -> bool

Check if the symbol table is empty

pub fn iter(&self) -> impl Iterator<Item = &Symbol>

Get an iterator over all symbols

pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut Symbol>

Iterate over all symbols with mutable access

pub fn get_main(&self) -> Option<&Symbol>

Get the main function

pub fn current_symbol(&self) -> Option<&Symbol>

Get the current symbol being processed (last added)

pub fn current_symbol_mut(&mut self) -> Option<&mut Symbol>

Get a mutable reference to the current symbol